+++ macOS/PrivateFrameworks/PhotosUICore.framework/PXDebugValueList.h
@@ -21,7 +21,7 @@
 - (void)addValueWithLabel:(id)arg1 highlightedScore:(double)arg2;
 - (void)addValueWithLabel:(id)arg1 stringValue:(id)arg2;
 - (void)addValueWithLabel:(id)arg1 stringValue:(id)arg2 highlightStyle:(unsigned long long)arg3;
-- (void)addValueWithLabel:(id)arg1 integerValue:(double)arg2;
+- (void)addValueWithLabel:(id)arg1 integerValue:(long long)arg2;
 - (void)addValueWithLabel:(id)arg1 doubleValue:(double)arg2;
 - (void)addValueWithLabel:(id)arg1 boolValue:(BOOL)arg2;
 - (void)addValueWithLabel:(id)arg1 boolValue:(BOOL)arg2 positiveValue:(BOOL)arg3 positiveHighlighted:(BOOL)arg4 negativeHighlighted:(BOOL)arg5;
