+++ macOS/iOSSupport/PrivateFrameworks/PhotosUICore.framework/PXFileBackedAsset.h
@@ -72,7 +72,7 @@
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, nonatomic) double duration;
 @property(readonly) unsigned long long hash;
-@property(readonly, nonatomic) double hdrGain;
+@property(readonly, nonatomic) float hdrGain;
 @property(readonly, nonatomic) NSDate *importDate;
 @property(readonly, nonatomic) NSString *localizedGeoDescription;
 @property(readonly) Class superclass;
